requirements.txt
Django>=5.2
requests
beautifulsoup4

djangorestframework
drf-spectacular

sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10

#DEBUG MODE.
debugpy>=1.6.0

selenium==4.16.0